exception handling and package rename
1.exception handling
2.package rename
