--- chap-1/spring/spring-mysql/src/main/java/com/klab/repository/TodoRepository.java
@@ -1,8 +1,0 @@
-package com.klab.repository;
-
-import org.springframework.data.repository.CrudRepository;
-
-import com.klab.entity.Todo;
-
-public interface TodoRepository extends CrudRepository<Todo,Long> {
-}
