+++ chap-1/spring/spring-mysql/src/main/java/com/klab/todo/TodoApplication.java
@@ -1,4 +1,4 @@
-package com.klab;
+package com.klab.todo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
