+++ chap-1/spring/spring-mysql/src/main/java/com/klab/todo/config/SwaggerConfig.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.klab.config;
+package com.klab.todo.config;
 
 import java.util.List;
