+++ chap-1/spring/spring-mysql/src/main/java/com/klab/todo/exception/TodoException.java
@@ -1,0 +1,14 @@
+package com.klab.todo.exception;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class TodoException {
+
+    private String message;
+    private String requestedURI;
+
+ 
+}
