+++ chap-1/spring/spring-mysql/src/main/java/com/klab/todo/repository/TodoRepository.java
@@ -1,0 +1,8 @@
+package com.klab.todo.repository;
+
+import org.springframework.data.repository.CrudRepository;
+
+import com.klab.todo.entity.Todo;
+
+public interface TodoRepository extends CrudRepository<Todo,Long> {
+}
